+++ repositories/web-application-stack/images/web-client/bin/entrypoint.bash
@@ -10,7 +10,7 @@ set -e
 echo "$1" > /run/docker-entrypoint-command.txt
 case "$1" in
     '--start-http-server')
-        substitute_environment_variables < /etc/opt/app/env.js.template > /opt/app/public/env.js
+        substitute_environment_variables < /etc/opt/app/config.json.template > /opt/app/public/config.json
         substitute_environment_variables < /etc/opt/app/index.html.template > /opt/app/public/index.html
 
         substitute_environment_variables < /etc/nginx/nginx.conf.template > /etc/nginx/nginx.conf